Add platform specific files.
