+++ input/src/event/platforms/mod.rs
@@ -1,0 +1,4 @@
+#[cfg(target_os="windows")]
+mod windows;
+#[cfg(target_os="linux")]
+mod linux;
