+++ input/src/lib.rs
@@ -1,10 +1,12 @@
 mod device_id;
 mod event;
 mod event_manager;
-mod event_reader;
-mod event_writer;
+mod platforms;
 mod glue;
 
 pub use event::{Axis, Button, Direction, Event, Key, KeyKind};
 pub use event_manager::EventManager;
-pub use event_writer::EventWriter;
+
+#[cfg(target_os="linux")]
+pub use platforms::linux::event_writer::EventWriter;
+pub use platforms::linux::event_reader::EventReader;
