+++ input/src/platforms/mod.rs
@@ -1,0 +1,2 @@
+pub mod linux;
+pub mod windows;
